requirements.txt
albucore==0.0.13
albumentations==1.4.10
annotated-types==0.7.0
anyio==4.8.0
astor==0.8.1
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
exceptiongroup==1.2.2
filelock==3.16.1
fire==0.7.0
Flask==3.1.0
Flask-Mail==0.10.0
fonttools==4.55.3
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
imgaug==0.4.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
lap==0.5.12
lazy_loader==0.4
lmdb==1.6.2
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
mysql-connector==2.2.9
mysql-connector-python==9.2.0
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
packaging==24.2
paddleocr==2.9.1
paddlepaddle==2.6.2
pandas==2.2.3
pillow==11.1.0
protobuf==3.20.2
psutil==6.1.1
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pydantic==2.10.5
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.11.0
requests==2.32.3
scikit-image==0.25.0
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
shapely==2.0.6
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.1
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2025.1.10
tomli==2.2.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.65
ultralytics-thop==2.0.14
urllib3==2.3.0
Werkzeug==3.1.3
